Update README.md
Updated the readme with the correct links and locations to the asset picture, and the contributors part etc.

Updating Readme and gradle setup for builds and intelij workspace.

Updated Gradle setup to work correctly with old Gradle 4.4.1 with the requirement of Lombok.

Updated the readme main for the wargames page, keeping the old readme and description in tact for viewers.

Updated the banner to match the name of the repository etc.

Updating Readme and gradle setup for builds and intelij workspace.

Updated Gradle setup to work correctly with old Gradle 4.4.1 with the requirement of Lombok.

Updated the readme main for the wargames page, keeping the old readme and description in tact for viewers.

Updated the banner to match the name of the repository etc.

Merge branch 'main' of https://github.com/Wargames-Development/MCHELI-R-WDG-Edition

Delete MCHELI-R-Banner.png.png

Updates the readme file with the correct link to our compiled source for private viewing within the team.