Update the MCH__MOD's modid data section + implement method for mod version mismatch kicks.
Updated the name and version of the mod in the modid data. Also implemented a checkModVersions method to avoid mod version mismatch that we believe keeps happening that will now ensure parity between the local users version and servers version based on the mcmod.info file's versioning to make it simple.

Therefore when you want a release & EVERY release you need to update that mcmod.info file to be correct for the versioning and now replace ALL of the required files not just individual class files

Updates versioning for this fix