Make scoreboard gui only op available
make the scoreboard gui which is inherently broken on our server due to the plugin scoreboard system that we use adding false teams that clutter, available to op only (extra gated).

Might eventually see if i can remove clutter or functionality or remove entirely since the plan is to have it managed by clowder only and be fully automatic. Might turn into a clowder faction management sort of system... Currently unsure but gating to ops fixes a lot of the issues anyway and avoids player accidental screwups of the automation in yRadar.